impl RbxWriteInterleaved for Vec<u8>
